+++ src/App.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+
+import logo from './lib/logo.svg';
+
+import './css/App.css';
 
 function App() {
   return (
