+++ src/index.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+
 import App from './App';
-import * as serviceWorker from './serviceWorker';
+
+import * as serviceWorker from './lib/serviceWorker';
+
+import './css/index.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
